Renamed search_endpoint.py to index_operations.py
